requirements.txt
asgiref==3.8.1
attrs==24.2.0
backports.zoneinfo==0.2.1
Django==4.2.16
djangorestframework==3.15.2
drf-spectacular==0.27.2
importlib-resources==6.4.5
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
pkgutil-resolve-name==1.3.10
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
typing-extensions==4.12.2
uritemplate==4.1.1
zipp==3.20.2
